rl-gripper/requirements.txt
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
matplotlib==3.8.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
Pillow==10.1.0
pybullet==3.2.6
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-qt-binding==1.1.2
pytz==2024.1
qt-dotgraph==2.2.3
stable_baselines3==2.3.0
statistics-msgs==1.2.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.1.2
torchvision==0.16.2
